Added modularize js code
