+++ index.html
@@ -78,6 +78,6 @@
          <p>Copyright @Omar_Sadaf <span id="footer-year"></span></p>
         </div>
       </footer>
-      <script src="assets/js/main.js"></script>
+      <script type="module" src="./assets/js/main.js"></script>
     </body>
 </html>
